[GP-8] extract components
